+++ app/src/main/java/com/example/comusenias/presentation/component/home/ContentCardGame.kt
@@ -4,6 +4,7 @@ import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.animateIntAsState
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -21,8 +22,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
 import com.example.comusenias.R
 import com.example.comusenias.constants.TestTag
+import com.example.comusenias.presentation.navigation.AppScreen
 import com.example.comusenias.presentation.ui.theme.SIZE12
 import com.example.comusenias.presentation.ui.theme.blackColorApp
 import com.example.comusenias.presentation.ui.theme.cardGray
@@ -40,7 +43,7 @@ enum class StatusGame {
 }
 
 @Composable
-fun ContentCardGame(status: StatusGame ) {
+fun ContentCardGame(status: StatusGame, navController: NavController ) {
     val currentStatus by remember { mutableStateOf(status) }
 
     val lineColor: Color
@@ -84,6 +87,7 @@ fun ContentCardGame(status: StatusGame ) {
             .fillMaxWidth()
             .blur(radius = blurAnimate.value)
             .background(Color.White)
+            .clickable { navController.navigate(AppScreen.LearnSignScreen.route) }
             .testTag(TestTag.TAG_CONTENT_CARD_GAME + currentStatus.name)
     ) {
         CardGame(
@@ -109,16 +113,4 @@ fun ContentCardGame(status: StatusGame ) {
             )
         }
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewContentCard(){
-    Column {
-        ContentCardGame(status = StatusGame.COMPLETED)
-        ContentCardGame(status = StatusGame.IN_PROGRESS)
-        ContentCardGame(status = StatusGame.BLOCKED)
-    }
-
-
 }
