+++ app/src/main/java/com/example/comusenias/presentation/screen/login/LoginScreen.kt
@@ -33,7 +33,7 @@ fun LoginScreen(
         modifier = modifier
             .fillMaxSize()
             .testTag(TestTag.TAG_LOGIN_SCREEN),
-        ) {
+    ) {
         Column(
             modifier = Modifier
                 .fillMaxSize()
