+++ app/src/main/java/com/example/comusenias/presentation/ui/theme/Size.kt
@@ -33,13 +33,16 @@ const val size50: Int = 50
 const val SIZE60: Int = 60
 const val SIZE64: Int = 64
 const val SIZE70: Int = 70
+const val SIZE73: Int = 73
 const val SIZE80: Int = 80
+const val SIZE84: Int = 84
 const val size95: Int = 95
 const val SIZE90: Int = 90
 const val SIZE100: Int = 100
 const val size130: Int = 130
 const val size150: Int = 150
 const val SIZE200: Int = 200
+const val SIZE220: Int = 220
 const val SIZE250: Int = 250
 const val SIZE300: Int = 300
 const val SIZE390: Int = 390
